[refactor] Extract response widget to new file
